pyproject.toml
[tool.poetry]
name = "django-sample"
version = "0.1.0"
description = ""
authors = ["aiden <aiden@safuventures.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
django-extensions = "^3.1.5"
django-environ = "^0.8.1"
django-filter = "^21.1"
django-activity-stream = "^1.4.0"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
dj-rest-auth = "^2.2.4"
django-allauth = "^0.50.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipdb = "^0.13.9"
ipython = "^8.3.0"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
